+++ client/src/features/facets/DateInput.tsx
@@ -4,6 +4,7 @@
 import React, { useRef, useState, useEffect } from 'react'
 import { useLocation, useNavigate, useParams } from 'react-router-dom'
 import styled from 'styled-components'
+import { Col, Row } from 'react-bootstrap'
 
 import { addLastSelectedFacet } from '../../redux/slices/facetsSlice'
 import { useAppDispatch } from '../../app/hooks'
@@ -16,9 +17,15 @@ import { ResultsTab } from '../../types/ResultsTab'
 import { pushClientEvent } from '../../lib/pushClientEvent'
 import { IFacetsPagination } from '../../types/IFacets'
 import { useGetFacetsSearchQuery } from '../../redux/api/ml_facets_api'
-import { getYearsFromFacetValues } from '../../lib/facets/dateParser'
+import {
+  getDateInputPlaceholder,
+  getEra,
+  getTimestampsFromFacetValues,
+} from '../../lib/facets/dateParser'
+import { timePeriod } from '../../config/advancedSearch/inputTypes'
 
 import DateSlider from './DateSlider'
+import FacetDropdown from './Dropdown'
 
 interface IFacets {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -63,37 +70,62 @@ const DateInput: React.FC<IFacets> = ({
   lastPage,
   autoFocus = false,
 }) => {
-  let yearOne = ''
-  let defaultLastYear = ''
+  const dispatch = useAppDispatch()
+
+  // Get initial values
+  let minDate = Date.now()
+  let defaultLastYear = Date.now()
+  let timestamps: Array<number> = []
   if (facetValues.requests.hasOwnProperty('call1')) {
-    const years = getYearsFromFacetValues(facetValues.requests.call1)
-    if (years.length > 0) {
-      yearOne = years[0].toString()
-      defaultLastYear = years[years.length - 1].toString()
+    timestamps = getTimestampsFromFacetValues(facetValues.requests.call1)
+    // timestamps = getTimestampsFromFacetValues([
+    //   {
+    //     id: '',
+    //     totalItems: 1,
+    //     type: 'OrderedCollection',
+    //     value: '-0148-01-01T00:00:00Z',
+    //   },
+    //   {
+    //     id: '',
+    //     totalItems: 1,
+    //     type: 'OrderedCollection',
+    //     value: '2000-01-01T00:00:00Z',
+    //   },
+    // ])
+    if (timestamps.length > 0) {
+      minDate = timestamps[0]
+      defaultLastYear = timestamps[timestamps.length - 1]
     }
   }
-  const dispatch = useAppDispatch()
-  const [earliest, setEarliest] = useState<string>(yearOne)
-  const [latest, setLatest] = useState<string>(defaultLastYear)
-  // This will only change upon retrieving the last date year
-  const [maxYear, setMaxYear] = useState<string>(defaultLastYear)
 
-  const { data, isSuccess } = useGetFacetsSearchQuery({
-    q: JSON.stringify(combinedQuery),
-    facets: {},
-    facetNames: facetName,
-    tab: currentTab,
-    page: lastPage,
-  })
+  const [earliest, setEarliest] = useState<number>(minDate)
+  const [latest, setLatest] = useState<number>(defaultLastYear)
+  const [earliestEra, setEarliestEra] = useState<string>(getEra(earliest))
+  const [latestEra, setLatestEra] = useState<string>(getEra(latest))
+  // This will only change upon retrieving the last date year
+  const [maxDate, setMaxDate] = useState<number>(defaultLastYear)
+  const disableEraDropdown = getEra(minDate) === getEra(maxDate)
+  const swapDates = disableEraDropdown && getEra(minDate) === 'bce'
+
+  const { data, isSuccess } = useGetFacetsSearchQuery(
+    {
+      q: JSON.stringify(combinedQuery),
+      facets: {},
+      facetNames: facetName,
+      tab: currentTab,
+      page: lastPage,
+      sort: 'desc',
+    },
+    {
+      skip: lastPage === 1,
+    },
+  )
 
   const navigate = useNavigate()
   const { pathname, search } = useLocation()
   const { tab } = useParams<keyof ResultsTab>() as ResultsTab
   const paramPrefix = searchScope[tab].slice(0, 1)
 
-  // const years = getYearsFromFacetValues(facetValues)
-
-  const [redirect, setRedirect] = useState(false)
   const earliestRef = useRef<HTMLInputElement>(null)
   const latestRef = useRef<HTMLInputElement>(null)
 
@@ -103,34 +135,125 @@ const DateInput: React.FC<IFacets> = ({
   useEffect(() => {
     if (isSuccess && data) {
       const { orderedItems } = data as ISearchResults
-      const yearsOfLastPage = getYearsFromFacetValues(orderedItems)
-      const yearsOfLastPageLength = yearsOfLastPage.length
-      const lastYearOfRange =
-        yearsOfLastPage[yearsOfLastPageLength - 1].toString()
-      setLatest(lastYearOfRange)
-      setMaxYear(lastYearOfRange)
+      const lastPageDates = getTimestampsFromFacetValues(orderedItems)
+      const datesOfLastPageLength = lastPageDates.length
+      const lastDateOfRange = lastPageDates[datesOfLastPageLength - 1]
+      setLatest(lastDateOfRange)
+      setLatestEra(getEra(lastDateOfRange))
+      setMaxDate(lastDateOfRange)
+      // if (swapDates) {
+      //   setLatest(earliest)
+      //   setEarliest(lastDateOfRange)
+      // }
     }
-  }, [data, isSuccess])
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [data, isSuccess, minDate])
 
-  const handleEarliestInputChange = (value: string): void => {
-    setEarliest(value)
+  const handleEarliestInputChange = (value: string | number): void => {
+    const date = new Date(value)
+    // The input comes from the date picker if it is a string
+    if (typeof value === 'string') {
+      // check if it is bce or not
+      if (earliestEra === 'bce') {
+        let year = date.getUTCFullYear()
+        year = -year
+        // Will need to add logic for adding date if user passes over BCE into CE and vice versa
+        date.setFullYear(year)
+        setEarliest(date.getTime())
+      } else {
+        setEarliest(date.getTime())
+      }
+    } else {
+      const year = date.getUTCFullYear()
+      if (earliestEra === 'bce' && year > 1) {
+        setEarliestEra('ce')
+      }
+      if (earliestEra === 'ce' && year < 1) {
+        setEarliestEra('bce')
+      }
+      setEarliest(value)
+    }
   }
 
-  const handleLatestInputChange = (value: string): void => {
-    setLatest(value)
+  const handleLatestInputChange = (value: string | number): void => {
+    const date = new Date(value)
+    // The input comes from the date picker if it is a string
+    if (typeof value === 'string') {
+      // check if it is bce or not
+      if (earliestEra === 'bce') {
+        let year = date.getUTCFullYear()
+        year = -year
+        // Will need to add logic for adding date if user passes over BCE into CE and vice versa
+        date.setFullYear(year)
+        setLatest(date.getTime())
+      } else {
+        setLatest(date.getTime())
+      }
+    } else {
+      const year = date.getUTCFullYear()
+      if (latestEra === 'bce' && year > 1) {
+        setLatestEra('ce')
+      }
+      if (latestEra === 'ce' && year < 1) {
+        setLatestEra('bce')
+      }
+      setLatest(value)
+    }
+  }
+
+  const handleEarliestEraChange = (selected: string): void => {
+    const date = new Date(earliest)
+    let year = date.getUTCFullYear()
+    if (year < 1) {
+      if (selected === 'ce') {
+        year = Math.abs(year)
+        date.setUTCFullYear(year)
+      }
+    }
+
+    if (year > 1) {
+      if (selected === 'bce') {
+        year = -year
+        date.setUTCFullYear(year)
+      }
+    }
+    setEarliest(date.getTime())
+    setEarliestEra(selected)
+    // will need to set earliest too
+  }
+
+  const handleLatestEraChange = (selected: string): void => {
+    const date = new Date(earliest)
+    let year = date.getUTCFullYear()
+    if (year < 1) {
+      if (selected === 'ce') {
+        year = Math.abs(year)
+        date.setUTCFullYear(year)
+      }
+    }
+
+    if (year > 1) {
+      if (selected === 'bce') {
+        year = -year
+        date.setUTCFullYear(year)
+      }
+    }
+    setLatest(date.getTime())
+    setLatestEra(selected)
+    // will need to set latest too
   }
 
   const submitHandler = (event: React.FormEvent<HTMLFormElement>): void => {
     event.preventDefault()
 
     if (earliest === null) {
-      const earliestYear = yearOne
-      setEarliest(earliestYear.toString())
+      const earliestYear = minDate
+      setEarliest(earliestYear)
     }
 
     if (latest === null) {
-      const latestYear = maxYear
-      setLatest(latestYear.toString())
+      const latestYear = maxDate
+      setLatest(latestYear)
     }
 
     dispatch(addLastSelectedFacet({ facetName: facetSection, facetUri: '' }))
@@ -162,6 +285,9 @@ const DateInput: React.FC<IFacets> = ({
       // TODO: get rid of keyof typeof
       searchTerms[facetSection as keyof typeof searchTerms]
     const criteriaKeys = Object.keys(query)
+    // Convert to ISO date
+    const earliestIsoDate = new Date(earliest)
+    const latestIsoDate = new Date(latest)
     if (criteriaKeys.includes('AND')) {
       const { AND } = query
       addFacetToArray(AND)
@@ -174,8 +300,8 @@ const DateInput: React.FC<IFacets> = ({
     return query
 
     function addFacetToArray(array: Array<ICriteria>): void {
-      const min = { [searchTermName]: earliest, _comp: '>=' }
-      const max = { [searchTermName]: latest, _comp: '<=' }
+      const min = { [searchTermName]: earliestIsoDate, _comp: '>=' }
+      const max = { [searchTermName]: latestIsoDate, _comp: '<=' }
       array.push(min, max)
       // TODO: uncomment when ML estimates are fixed
       // array.push({ [searchTermName]: { start: earliest, end: latest } })
@@ -187,49 +313,103 @@ const DateInput: React.FC<IFacets> = ({
       <form className="w-100" onSubmit={submitHandler}>
         <div className="input-group d-block">
           <DateSlider
-            min={parseInt(yearOne, 10)}
-            max={parseInt(maxYear, 10)}
+            min={swapDates ? maxDate : minDate}
+            max={swapDates ? minDate : maxDate}
             earliestVal={earliest}
             latestVal={latest}
             onEarliestChange={handleEarliestInputChange}
             onLatestChange={handleLatestInputChange}
           />
           <div className="d-flex justify-content-between pt-4">
-            <div className="justify-content-start">
-              <label htmlFor={earliestDateId} className="d-none">
-                Earliest Date
-              </label>
-              <StyledInput
-                id={earliestDateId}
-                type="number"
-                className="form-control"
-                placeholder={yearOne.toString()}
-                onChange={(e) =>
-                  handleEarliestInputChange(e.currentTarget.value)
-                }
-                min={parseInt(yearOne, 10)}
-                max={parseInt(maxYear, 10)}
-                ref={earliestRef}
-                value={earliest}
-                autoFocus={autoFocus}
-              />
-            </div>
-            <div className="justify-content-end">
-              <label htmlFor={latestDateId} className="d-none">
-                Latest Date
-              </label>
-              <StyledInput
-                id={latestDateId}
-                type="number"
-                className="form-control"
-                placeholder={maxYear}
-                onChange={(e) => handleLatestInputChange(e.currentTarget.value)}
-                min={parseInt(yearOne, 10)}
-                max={parseInt(maxYear, 10)}
-                ref={latestRef}
-                value={latest}
-              />
-            </div>
+            <Row>
+              <Col xs="12">
+                <label htmlFor={earliestDateId}>Earliest Date</label>
+              </Col>
+              <Col>
+                <StyledInput
+                  id={earliestDateId}
+                  type="date"
+                  className="form-control"
+                  placeholder={getDateInputPlaceholder(earliest)}
+                  onChange={(e) =>
+                    handleEarliestInputChange(e.currentTarget.value)
+                  }
+                  // min={
+                  //   getEra(minDate) === 'bce'
+                  //     ? undefined
+                  //     : getDateInputPlaceholder(earliest)
+                  // }
+                  // max={
+                  //   getEra(minDate) === 'bce'
+                  //     ? undefined
+                  //     : getDateInputPlaceholder(maxDate)
+                  // }
+                  ref={earliestRef}
+                  value={getDateInputPlaceholder(earliest)}
+                  autoFocus={autoFocus}
+                />
+              </Col>
+              <Col>
+                <label htmlFor="earliest-date-period" hidden>
+                  Select CE or BCE
+                </label>
+                <FacetDropdown
+                  options={timePeriod}
+                  handleChange={handleEarliestEraChange}
+                  className="earliestYearEraSelection me-2"
+                  dropdownHeaderText="Select the era"
+                  ariaLabel="Select the era BC or BCE"
+                  selected={earliestEra}
+                  id="earliest-date-period"
+                  disabled={disableEraDropdown}
+                />
+              </Col>
+            </Row>
+          </div>
+          <div className="d-flex justify-content-between pt-4">
+            <Row>
+              <Col xs="12">
+                <label htmlFor={latestDateId}>Latest Date</label>
+              </Col>
+              <Col>
+                <StyledInput
+                  id={latestDateId}
+                  type="date"
+                  className="form-control"
+                  placeholder={getDateInputPlaceholder(maxDate)}
+                  onChange={(e) =>
+                    handleLatestInputChange(e.currentTarget.value)
+                  }
+                  // min={
+                  //   getEra(maxDate) === 'bce'
+                  //     ? undefined
+                  //     : getDateInputPlaceholder(earliest)
+                  // }
+                  // max={
+                  //   getEra(maxDate) === 'bce'
+                  //     ? undefined
+                  //     : getDateInputPlaceholder(maxDate)
+                  // }
+                  ref={latestRef}
+                  value={getDateInputPlaceholder(latest)}
+                />
+              </Col>
+              <Col>
+                <label htmlFor="latest-date-period" hidden>
+                  Select CE or BCE
+                </label>
+                <FacetDropdown
+                  options={timePeriod}
+                  handleChange={handleLatestEraChange}
+                  className="latestYearEraSelection me-2"
+                  dropdownHeaderText="Select the era"
+                  ariaLabel="Select the era BC or BCE"
+                  selected={latestEra}
+                  id="latest-date-period"
+                  disabled={disableEraDropdown}
+                />
+              </Col>
+            </Row>
           </div>
         </div>
         <div className="d-flex justify-content-end">